part 2 is hard...

+++ Day11/AddOperation.cpp
@@ -1,7 +1,7 @@
 #include "AddOperation.hpp"
 
 
-AddOperation::AddOperation(int value) : Operation(value)
+AddOperation::AddOperation(unsigned long long int value) : Operation(value)
 {
 }
 
@@ -9,7 +9,7 @@ AddOperation::~AddOperation()
 {
 }
 
-int AddOperation::PerformOperation(int input)
+unsigned long long int AddOperation::PerformOperation(unsigned long long int input)
 {
     return m_value + input;
 }
