+++ Day11/MultiplyOperation.cpp
@@ -1,7 +1,7 @@
 #include "MultiplyOperation.hpp"
 
 
-MultiplyOperation::MultiplyOperation(int value) : Operation(value)
+MultiplyOperation::MultiplyOperation(unsigned long long int value) : Operation(value)
 {
 }
 
@@ -9,7 +9,7 @@ MultiplyOperation::~MultiplyOperation()
 {
 }
 
-int MultiplyOperation::PerformOperation(int input)
+unsigned long long int MultiplyOperation::PerformOperation(unsigned long long int input)
 {
     return m_value * input;
 }
